Make `wallabag-request-and-insert-entries` handle large query sizes
This function was setting the perPage API query parameter to the number of
requested entries. This causes the API call to fail when the query number is
large. This version breaks up the queries into batches of 30. (the default
according to the API docs:
https://app.wallabag.it/api/doc#get--api-entries.%7B_format%7D )

It exposes one additional optional paramter (the page number of results
requested) but this does not affect any of the other calls to this function.